feat(api): :sparkles: file meta endpoint, file list size, bug fixes
A new feature was added which allows you to see a files metadata such as hashes, file size and more. You can get the meta by providing a `meta` query to the url. The directory listing api now also includes the file size. A bug was fixed which was showing a file as wrongfully inexinstent if a query parameter was specified. The logger got an upgrade.